LPS-202017 Move to the right package
