+++ modules/apps/ai-creator/ai-creator-openai-api/src/main/java/com/liferay/ai/creator/openai/display/context/factory/AICreatorOpenAIMenuItemFactory.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.ai.creator.openai.display.context.util;
+package com.liferay.ai.creator.openai.display.context.factory;
 
 import com.liferay.portal.kernel.servlet.taglib.ui.MenuItem;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
